Funciona correctamente: postService

+++ api/src/models/servicio.js
@@ -14,7 +14,7 @@ module.exports = (sequelize) => {
             unique: true
         },
         price: {
-            type: DataTypes.DECIMAL(3,2),
+            type: DataTypes.BIGINT,  // Permites números enteros entre -9,435,465,567,345,789,765 y 9,352,453,567,234,435,898.
             allowNull: false
         }
     }, { timestamps: false })
